refactor: split app.py routes into dedicated pages module for improved organization
